third_party/blink/web_tests/external/wpt/webrtc/RTCRtpReceiver-getSynchronizationSources.https-expected.txt

This is a testharness.js-based test.
[FAIL] [audio-only] RTCRtpSynchronizationSource.voiceActivityFlag is a boolean
  assert_equals: expected "boolean" but got "undefined"
Harness: the test ran to completion.
